fix: fixed datatestIds

+++ back-end/src/database/models/products.js
@@ -2,7 +2,7 @@ module.exports = (sequelize, DataTypes) => {
   const Products = sequelize.define('Products', {
     id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
     name: DataTypes.STRING(100),
-    price: DataTypes.DECIMAL(4,2),
+    price: DataTypes.FLOAT(4,2),
     url_image: DataTypes.STRING(200),
   },
   {
